langGraph/Minimal-Tutor-AI/requirements.txt
huggingface_hub
langgraph
langchain_openai
langchain_community
python-dotenv
langchain
duckduckgo-search
wikipedia
langchain-huggingface
faiss-cpu
tavily-python
psycopg2